Remove code for the particle photon board.
Both boards I own are now fried.

The plan is to replace the photon board with an esp32. Ideally the code would
compile for both arduino and esp32 but if this proves to be too difficult then
I will keep a fallback branch that has code that can be compile to run on
arduino.