Change GPIO pins
Connected GPIO 12 to the relay module was causing the USB serial
communication to fail while powering the unit via an external power
source. This may be because the relay module was pulling down (or up?) the
pin. Using a different pin solves the issue.

Also set the values used for the ADC pins.