In which I eat my words
The Stepper library should work, but it doesn't for some reason. One of
the symptoms is that in3 is never set to high - despite in entering the
case where this should happen.

The library does have some nice ideas though, using checking diff in
time rather than using delay and using the remainder to select case, so
I've borrowed those and added it to the Baffel code.